In an attempt to be able to understand the entire library inside-out and keep it simple (meaning targeted just for MMURTL) I hope to be able to actually get this working.
This builds off the CM32 code as well as my MMURTL system call
library for gcc and uses Plauger's "The Standard C Library" as
more of a guide and less of a specific implementation.

Taking what I have learned from trying to port newlib and
Plauger's code, I hope to make this something that anyone who
wants to play with MMURTL can understand and use.