Discovered issue with pointers to functions when compiling for MMURTL with gcc.  The loader script builds code to addresses relative to 0x0 + the MMURTL .RUN header (0x37).  Problem is, when loaded, the address is up in the 0x40000000 + stack - header + address!
Added two variables to make the loader script more user-friendly
and these values can be used in your code if needed.  This WAS needed
to get the signal function to work.

These values are _stack and _header for the size of your stack
and the size of the .RUN header.  There is also a variable for where
your data segment begins and will need to be adjusted if you find
your text segment going above 0x1000 bytes or if you change the stack
size.  All of these variables are towards the top of the linker
script.